apps/backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentic-rag-backend"
version = "0.1.0"
description = "Agentic RAG Customer Service — Backend (FastAPI + DDD)"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]",
    "uvicorn[standard]",
    "pydantic-settings",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg",
    "redis",
    "qdrant-client",
    "dependency-injector",
    "langgraph",
    "langchain-core",
    "httpx",
    "python-jose[cryptography]>=3.5.0",
    "pypdf>=6.7.2",
    "python-docx>=1.2.0",
    "striprtf>=0.0.29",
    "langchain-text-splitters>=1.1.1",
    "structlog>=25.5.0",
    "bcrypt>=4.0",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-bdd>=8.1",
    "pytest-asyncio",
    "pytest-cov",
    "ruff",
    "mypy",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
pythonpath = "."
addopts = "-v -p no:asyncio"
bdd_features_base_dir = "tests/features/"
markers = [
    "integration: Integration tests requiring real DB (docker-compose)",
]
filterwarnings = [
    "ignore:The garbage collector is trying to clean up:sqlalchemy.exc.SAWarning",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "C", "B"]
ignore = ["B008"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
check_untyped_defs = true

[tool.coverage.run]
branch = true
source = ["src"]
parallel = false
omit = [
    "tests/*",
    "src/infrastructure/db/seed.py",
    "src/infrastructure/db/engine.py",
    "src/infrastructure/db/base.py",
    "src/infrastructure/db/models/*",
    "src/infrastructure/db/repositories/*",
    "src/infrastructure/db/health_repository.py",
    "src/infrastructure/logging/*",
    "src/infrastructure/auth/*",
    "src/infrastructure/embedding/openai_embedding_service.py",
    "src/infrastructure/llm/anthropic_llm_service.py",
    "src/infrastructure/llm/openai_llm_service.py",
    "src/infrastructure/qdrant/*",
    "src/infrastructure/sentiment/*",
    "src/infrastructure/langgraph/agent_graph.py",
    "src/infrastructure/langgraph/langgraph_agent_service.py",
    "src/infrastructure/langgraph/tools.py",
    "src/infrastructure/tools/*",
    "src/infrastructure/line/*",
    "src/interfaces/*",
    "src/container.py",
    "src/main.py",
]

[tool.coverage.report]
fail_under = 70
show_missing = true
precision = 2
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.",
]
